Flying Chicken Game
This is a simple game that you can play on keyboard. In the game there is a flying chicken that you can control it and also this chicken can shoot eggs to eliminate the attackers such as tigers in the game. You can go right, left, upwards, downwards and also you can open parachute to avoid tigers and arrange your speed. The goal in the game is to have the highest score.